Fixes unit tests of existing features (#134)
# Description

This PR checks that all the existing scripts in the codebase work as
expected. The last few merges led to the deletion of some files that
were no longer available to import in the unit tests. Also renamed files
in the test to follow the following convention:

* `test_*.py` : A script that tests a feature using unittest fixture
* `check_*.py`: A script that runs the feature but doesn't necessarily
put them as a unittest fixture.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file