Puts the AppLauncher tutorial later in the series (#341)
# Description

Currently, the very first tutorial for users is the app launcher. This
might be a bit overwhelming to start with. This MR iterates over the app
launcher tutorial and puts it later in the part 1 of the tutorials.

## Type of change

- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there