Simplifies running of scripts and unit-tests (#448)
# Description

With Isaac Sim 2023.1, it seems we no longer need to have try-catch
blocks in the scripts. This allows us to simplify all the scripts and
test runners. Along with that, it enables VSCode debugger support which
is crucial for many developers.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have run all the tests with `./orbit.sh --test` and they pass
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there