Changes link names ordering in articulation to follow PhysX (#454)
# Description

This MR fixes an issue with the `Articulation` class. The indexing of
body names inside the articulation view is not the same as in the body
view. This affects any quantity of bodies (such as Jacobians) you try to
get directly from the articulation view.

The MR changes the `Articulation` class to follow the internal PhysX
articulation-view ordering of the link names. It deals internally with
re-ordering the articulation-link indices to body-view indices where
needed.

With this change, users no longer need to work with `body_physx_view` as
it will lead to unexpected behaviors. The MR removes this property from
the `RigidObject` and `Articulation` class.

Fixes #453

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have run all the tests with `./orbit.sh --test` and they pass
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there