Modifies container.sh and x11.yaml to automatically check for stale cookies (#558)
# Description

NOTE: This is a redo of
https://github.com/isaac-orbit/IsaacLab/pull/524, only rebased onto new
main and with changes to work with Isaac Lab.

Copied from other PR:
Modifies container.sh and x11.yaml to automatically check if a
tmp.***.xauth file has a stale MIT-MAGIC-COOKIE-1, and refreshes it if
so.

In the instance where an xsession ends after a `tmp.***.xauth` file has
been created, but the system is not rebooted (thus deleting
`tmp.***.xauth`) then attempting to use the container causes errors.

This is a feature to correct for this upon `container.sh enter` by:
1. Mounting an entire `/tmp` subdirectory so that we can swap files
2. Checking for stale cookies by comparing against the current user's
cookie
3. If stale, replacing the old `tmp.***.xauth` with a new
`tmp.***.xauth` of the same name but a new cookie, in order to preserve
functionality of envars.

## Type of change

<!-- As you go through the list, delete the ones that are not
applicable. -->

- New feature (non-breaking change which adds functionality)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have run all the tests with `./orbit.sh --test` and they pass
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

Signed-off-by: James Smith <142246516+jsmith-bdai@users.noreply.github.com>
Co-authored-by: Hunter Hansen <50837800+hhansen-bdai@users.noreply.github.com>