Disables default viewport when headless but cameras are enabled (#851)
# Description

When cameras are enabled, a default render product always gets created
for the viewport, which brings additional overhead for rendering the
viewport at each step. However, when running in headless mode, we may
not need to render the viewport if we are not recording or
livestreaming.

This change disables viewport rendering in the headless cases where
recording and livestreaming are not required.

## Type of change

- Bug fix (non-breaking change which fixes an issue)
- New feature (non-breaking change which adds functionality)


## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->