Adds function to define camera configs through intrinsic matrix (#617)
# Description

This PR adds the possibility of initializing cameras (both Raycaster
Cameras and USD Cameras) with the intrinsic matrix instead of using the
aperture parameters. The intrinsic matrix is defined in the pattern
config and the pinhole cameras spawn config, respectively.

Moreover, it fixes the bug that the vertical aperture is not adjusted
for the USD camera case (it will always stay at the default value, even
if the horizontal aperture is set). The default is squared pixels;
however, it also allows the setting of other values.

Fixes https://github.com/isaac-orbit/IsaacLab/issues/226

## Type of change

- New feature (non-breaking change which adds functionality)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there