Fixes Jacobian joint indices for floating base articulations (#1033)
# Description

Adds `6` to the joint indices when indexing the Jacobian tensor in
floating base systems. This is required because, in floating base
articulations, the first 6 elements of the last dimensions are w.r.t
to the root pose, not the joints.

Fixes #1032

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there