Removes DLAA workaround for frame offset issue when rendering (#175)
Previously, small resolution renders that used DLSS for antialiasing
would sometimes display frame offset issues where the rendered frame is
not in sync with the current frame. To workaround this issue, we forced
the antialiasing mode to DLAA for small resolutions in both the Camera
and TiledCamra classes.

With recent changes in Isaac Sim 4.5, the frame offset issue no longer
appears, so this PR removes the workaround we've put in to avoid this
problem.

<!-- As you go through the list, delete the ones that are not
applicable. -->

- Bug fix (non-breaking change which fixes an issue)

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [ ] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->