Updates docs for segmentation and 50 series GPUs issues (#1943)
# Description

This is a documentation only change that adds instructions for updating
pytorch to the nightly build when running with 50 series GPUs to avoid
CUDA issues. In addition, we added the v2.0.1 release notes and an extra
note for the v2.0.0 release for a breaking change in the semantic and
instance segmentation outputs when using tiled cameras.

Fixes #1888

## Type of change

- This change requires a documentation update

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./isaaclab.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

<!--
As you go through the checklist above, you can mark something as done by
putting an x character in it

For example,
- [x] I have done this task
- [ ] I have not done this task
-->

---------

Signed-off-by: Kelly Guo <kellyg@nvidia.com>